refactor: move hexstring_to_vec to higher-level utils
- The lower-level utils is feature-gated and can't be accessed from inside the higher-level Token object (which we need to check for a valid address).
- Wasn't sure if it's best to put it in this dir - or to move it into the protocol dir. I found it weird that this imports SimulationError from the protocol dir. Let me know if you have a preference.
