+++ src/evm/protocol/vm/utils.rs
@@ -320,36 +320,6 @@ pub fn load_erc20_bytecode() -> Result<Bytecode, FileError> {
     Ok(erc_20_bytecode)
 }
 
-/// Converts a hexadecimal string into a `Vec<u8>`.
-///
-/// This function accepts a hexadecimal string with or without the `0x` prefix. If the prefix
-/// is present, it is removed before decoding. The remaining string is expected to be a valid
-/// hexadecimal representation, otherwise an error is returned.
-///
-/// # Arguments
-///
-/// * `hexstring` - A string slice containing the hexadecimal string. It may optionally start with
-///   `0x`.
-///
-/// # Returns
-///
-/// * `Ok(Vec<u8>)` - A vector of bytes decoded from the hexadecimal string.
-/// * `Err(SimulationError)` - An error if the input string is not a valid hexadecimal
-///   representation.
-///
-/// # Errors
-///
-/// This function returns a `SimulationError::EncodingError` if:
-/// - The string contains invalid hexadecimal characters.
-/// - The string is empty or malformed.
-pub fn hexstring_to_vec(hexstring: &str) -> Result<Vec<u8>, SimulationError> {
-    let hexstring_no_prefix =
-        if let Some(stripped) = hexstring.strip_prefix("0x") { stripped } else { hexstring };
-    let bytes = hex::decode(hexstring_no_prefix)
-        .map_err(|_| SimulationError::FatalError(format!("Invalid hex string: {}", hexstring)))?;
-    Ok(bytes)
-}
-
 pub fn string_to_bytes32(pool_id: &str) -> Result<[u8; 32], SimulationError> {
     let pool_id_no_prefix =
         if let Some(stripped) = pool_id.strip_prefix("0x") { stripped } else { pool_id };
@@ -370,6 +340,7 @@ pub fn string_to_bytes32(pool_id: &str) -> Result<[u8; 32], SimulationError> {
 mod tests {
     use super::*;
 
+    use crate::utils::hexstring_to_vec;
     use dotenv::dotenv;
     use std::{fs::remove_file, io::Write};
     use tempfile::NamedTempFile;
