+++ src/lib.rs
@@ -22,3 +22,4 @@ pub mod evm;
 pub mod models;
 pub mod protocol;
 pub mod serde_helpers;
+mod utils;
